requirements.txt
discord.py
spotipy
yt_dlp
python-dotenv
psycopg2-binary
wavelink